requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bcrypt==4.1.3
beautifulsoup4==4.13.4
blinker==1.8.2
blis==0.7.11
botocore==1.42.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
datasets==3.6.0
dependencies==7.7.0
deprecation==2.1.0
dill==0.3.8
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
fastapi==0.110.1
fastapi-cli==0.0.7
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
frozenlist==1.8.0
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.39.1
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenery==4.2.2
greenlet==3.1.1
grpcio==1.66.0
grpcio-status==1.62.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.24.6
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jose==1.0.0
kiwisolver==1.4.7
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
motor==3.3.1
MouseInfo==0.1.3
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.10
numpy==2.4.0
oauthlib==3.3.0
ollama==0.5.1
openpyxl==3.1.5
orjson==3.10.18
packaging==24.1
pandas==2.3.3
passlib==1.7.4
pillow==11.0.0
pluggy==1.6.0
postgrest==1.1.1
preshed==3.0.9
propcache==0.3.2
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.5
pydantic-settings==2.2.1
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.5.0
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
PyYAML==6.0.2
qrcode==8.2
realtime==2.7.0
regex==2024.7.24
requests==2.32.5
requests-oauthlib==2.0.0
responses==0.25.8
rich==13.8.0
rich-toolkit==0.14.7
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.4.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==74.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.7
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.4.8
starlette==0.37.2
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tenacity==9.1.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.6.2
uvicorn==0.25.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.0.4
wikipedia==1.4.0
wrapt==1.16.0
xlrd==2.0.2
xmltodict==1.0.2
xxhash==3.5.0
yarl==1.22.0
